backend/requirements.txt

# Used by pip to install required python packages
# Usage: pip install -r requirements.txt

Flask>=0.11
gunicorn
Flask-Cors
langchain
langchain-community
langchain-google-genai
google-generativeai
python-dotenv
youtube-transcript-api